REMOVED: Some old/wrong/unneeded hackfixes, http://arcemu.org/forums/index.php?showtopic=23182 , thx Neo_mat FIXED: 2 warnings.

+++ src/arcemu-world/Player.cpp
@@ -11833,7 +11833,7 @@ void Player::Social_SendFriendList(uint32 flag)
 			data << uint8(0);
 
 		// online/offline flag
-		plr = objmgr.GetPlayer( itr->first );
+		plr = objmgr.GetPlayer((uint32)itr->first);
 		cache = objmgr.GetPlayerCache((uint32)itr->first);
 		if( plr != NULL )
 		{
